Move most breakpoint management logic to BreakpointManager.
